create actions folder and greeting file
