+++ src/reducers/greeting.js
@@ -1,10 +1,10 @@
-export const FETCH_GREETING = 'FETCH_GREETING';
+import { FETCH_GREETING } from '../actions/greeting';
 
 const initialState = {
   greeting: ''
 };
 
-export const reducer = (state = initialState, action) => {
+const greetingReducer = (state = initialState, action) => {
   switch (action.type) {
     case FETCH_GREETING:
       return { ...state, greeting: action.greeting.phrase };
@@ -13,23 +13,4 @@ export const reducer = (state = initialState, action) => {
   }
 };
 
-const handleErrors = (response) => {
-  if (!response.ok) {
-    throw Error(response.statusText);
-  }
-  return response;
-};
-
-export const setGreeting = greeting => ({
-  type: FETCH_GREETING,
-  greeting
-});
-
-export const fetchGreeting = () => (dispatch) => {
-  fetch('api/greetings')
-    .then(handleErrors)
-    .then(res => res.json())
-    .then((greeting) => {
-      dispatch(setGreeting(greeting.data[0]));
-    });
-};
+export default greetingReducer;
