+++ src/reducers/index.js
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
 import { routerReducer } from 'react-router-redux';
-import { reducer } from './greeting';
+import greetingReducer from './greeting';
 
 export default combineReducers({
   routing: routerReducer,
-  reducer
+  greetingReducer
 });
